requirements.txt
anyio==4.10.0
bio==1.6.2
biopython==1.85
biothings_client==0.4.1
certifi==2025.8.3
charset-normalizer==3.4.3
einops==0.8.1
exceptiongroup==1.3.0
fair-esm==2.0.0
filelock==3.19.1
fsspec==2025.7.0
gprofiler-official==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
mygene==3.2.2
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.1
platformdirs==4.3.8
pooch==1.8.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
